+++ NEWS.md
@@ -7,6 +7,7 @@
 - `tokenize_skip_ngrams` now supports stopwords (#31)
 - `tokenize_skip_ngrams` has been improved to generate unigrams and bigrams, according to the skip definition (#24)
 - Keyboard interrupt checks have been added to Rcpp-backed functions to enable users to terminate them before completion (#37)
+- New functions to count words, characters, and sentences without tokenization (#36).
 
 # tokenizers 0.1.4